+++ src/VO/PurchaseReceiptVO.java
@@ -109,7 +109,7 @@ public class PurchaseReceiptVO {
 		return totalPrice;
 	}
 
-	public void setTotalPrice(long totalPrice) {
+	public void setTotalPrice(double totalPrice) {
 		this.totalPrice = totalPrice;
 	}
